+++ src/memory/FastMemset.h
@@ -1,12 +1,11 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 // High-throughput memory fill with runtime SIMD dispatch on supported CPUs.
 // Falls back to a scalar implementation on unsupported architectures.
-void* fastMemset(void* destination, uint8_t value, size_t size);
+void* fastMemset(void* destination, uint8_t value, uint64_t size);
 
-inline void* fastZeroMem(void* destination, size_t size) {
+inline void* fastZeroMem(void* destination, uint64_t size) {
     return fastMemset(destination, 0, size);
 }
